requirements.txt
py-cord
numpy
python-dotenv
workers-kv-py
git+https://github.com/tomoemon/google_input.git@develop